+++ src/piglets/profiling/__init__.py
@@ -1,0 +1,5 @@
+from .profiler import Profiler as Profiler
+
+__all__ = [
+    "Profiler",
+]